Better sampling for the interpolation method.
It seems like there is no great difference between the
integration method (either with Newton-Cotes or Simpson)
and the interpolation method.

The latter needs a robust sampling, which may not be so simple
as implemented now, and Simpson is much slower than Newton-Cotes,
so we decided to continue using the original program.