Revert to the integration method (see comments!)
The precision problems we encountered with the integration method,
and also with the integration method, appear in a different guise
in the Bezier extraction method, as well.

The reason is that we have assumed that the cross derivative function
is C2, but in reality it is only C1, since it contains the first
derivative of the curve.

The combineBases() function still does not take care of this, but if we
manually set the combined basis s.t. all internal knots have degree-1
multiplicity, all methods work as expected.

The integration method is the simplest, and probably the fastest, so it
is chosen as the default.