Minor change to accomodate different inputs.
Target = Curve + CrossDer * Scaling + Curve' * Scaling

The original concept was cubic curve, cubic crossderivative, cubic
scaling, so it will be sextic (D3 + D3 * D3 + D2 * D3).

But this would lead to C1 continuity (because of the derivative of the
curve), so we are going to use quartic curves. This will also be sextic
(D4 + D3 * D3 + D3 * D3).

This changeset mainly fixes the program to work in this setting, i.e.,
taking the degree of the crossderivative (instead of the curve) to
compute the final degree etc.